[INT-204] Replace URL search calendar state with history state + TanStack Query (#176)
* [INT-204] Replace URL search calendar state with history state + TanStack Query

* Always fetch a month at a time

* Remove day from query key now that we fetch a month at a time

* Make filter non-optional

* Un-export getCalendarEvents
