+++ .eslintrc.json
@@ -2,7 +2,8 @@
   "extends": [
     "next/core-web-vitals",
     "prettier",
-    "plugin:storybook/recommended"
+    "plugin:storybook/recommended",
+    "plugin:@tanstack/query/recommended"
   ],
   "plugins": ["@typescript-eslint"],
   "rules": {
