+++ app/(authenticated)/calendar/helpers.ts
@@ -1,0 +1,13 @@
+import { QueryKey } from "@tanstack/react-query";
+
+export function calendarEventsQueryKey({
+  year,
+  month,
+  filter,
+}: {
+  year: number;
+  month: number;
+  filter: "all" | "mine" | "vacant";
+}) {
+  return ["fetchEvents", { year, month, filter }] as const satisfies QueryKey;
+}
