+++ package.json
@@ -98,6 +98,7 @@
     "@storybook/react": "^7.0.27",
     "@storybook/testing-library": "^0.0.14-next.2",
     "@tailwindcss/forms": "^0.5.3",
+    "@tanstack/eslint-plugin-query": "^5.59.7",
     "@types/ws": "^8.5.9",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "dotenv-cli": "^7.2.1",
